Initial commit of existing files
The How To MANTIS talks cover a variety of subjects, so of which have
accompanying example code or files. All of those files can be found in
this directory. This directory is also intended to function as an
interactive example for the git talk.

The contents of these directories are referenced by or used in the
following talks, and the general contents of each directory are as
described below.

* classes/
    - Talk name(s): classes.pptx
    - Contents: Python class example scripts

* dunders/
    - Talk name(s): dunders.pptx
    - Contents: Example Python dunder scripts

* povray/
    - Talk name(s): POV-Ray_basics.pptx
    - Contents: Example scripts shown on slides in the talk

* pytest/
    - Talk name(s): unittests.pptx
    - Contents: Example unit tests using the pytest module

* singularity/
    - Talk name(s): containers.pptx
    - Contents: definition file for an Arch Linux container

* unittest/
    - Talk name(s): unittests.pptx
    - Contents: Example unit tests using the unittest module

* vim/
    - Talk name(s): vim.pptx, sed.pptx
    - Contents: example file for editing practice, and a .vimrc example